refactor(blorktools): file handler deletion
