+++ packages/blorktools/src/asset_debugger/util/upload/file-upload-manager.js
@@ -482,3 +482,14 @@ function preventDefaults(e) {
     e.preventDefault();
     e.stopPropagation();
 }
+
+/**
+ * Format file size for display
+ * @param {number} bytes - The file size in bytes
+ * @returns {string} Formatted file size
+ */
+export function formatFileSize(bytes) {
+    if (bytes < 1024) return bytes + ' bytes';
+    else if (bytes < 1048576) return (bytes / 1024).toFixed(1) + ' KB';
+    else return (bytes / 1048576).toFixed(1) + ' MB';
+}
