+++ packages/blorktools/src/asset_debugger/util/upload/handlers/lighting-file-handler.js
@@ -1,8 +1,9 @@
-import { createClearButton, formatFileSize } from '../../../landing-page/file-handler';
 import { updateState } from '../../state/scene-state';
 import { processLightingFile } from '../../workers/worker-manager';
 import * as worldPanelModule from '../../../panels/world-panel/world-panel';
 import { hidePreviewLoading, showPreviewLoading } from '../../../loading-splash/preview-loading-splash';
+import { formatFileSize } from '../file-upload-manager';
+import { createClearButton } from '../../../landing-page/landing-page';
 
 /**
  * Handle lighting file upload
