+++ packages/blorktools/src/asset_debugger/util/upload/handlers/model/model-file-manager.js
@@ -1,6 +1,7 @@
-import { createClearButton, formatFileSize } from "../../../../landing-page/file-handler";
+import { createClearButton } from "../../../../landing-page/landing-page";
 import { hidePreviewLoading, showPreviewLoading } from "../../../../loading-splash/preview-loading-splash";
 import { updateState } from "../../../state/scene-state";
+import { formatFileSize } from "../../file-upload-manager";
 import { createGLBPreview } from "../../glb-preview-controller";
 import { processGLBFile } from "./glb-file-handler";
