+++ packages/blorktools/src/asset_debugger/util/upload/handlers/zip-file-handler.js
@@ -1,4 +1,4 @@
-import { formatFileSize } from "../../../landing-page/file-handler";
+import { formatFileSize } from "../file-upload-manager";
 
 /**
  * Handle ZIP file upload
